Update identity definition in TypesAsCategory
`identity a = (\x => x)` seems to be definitionally equivalent to `identity a = id`, but in practice the latter makes it difficult to use.

See the attempted definition for the constant Functor_Bool (Idris.Type -> Idris.Type) here: https://gitlab.com/snippets/1933362 that results in the hole `id = (\x => x)` which for some reason unknown to me seems impossible to fill. The proposed changed makes the definition possible.